refactor: 폴더 변경 및 서비스 클래스 추가
