+++ src/main/java/com/fairing/fairplay/qr/dto/scan/CheckOutRequestDto.java
@@ -1,0 +1,5 @@
+package com.fairing.fairplay.qr.dto.scan;
+
+public class CheckOutRequestDto {
+
+}
