+++ src/main/java/com/fairing/fairplay/qr/service/QrTicketEntryService.java
@@ -1,26 +1,21 @@
 package com.fairing.fairplay.qr.service;
 
 import com.fairing.fairplay.attendee.entity.Attendee;
-import com.fairing.fairplay.attendee.entity.AttendeeTypeCode;
-import com.fairing.fairplay.attendee.repository.AttendeeRepository;
-import com.fairing.fairplay.common.exception.CustomException;
 import com.fairing.fairplay.core.security.CustomUserDetails;
-import com.fairing.fairplay.qr.dto.CheckInRequestDto;
-import com.fairing.fairplay.qr.dto.CheckInResponseDto;
-import com.fairing.fairplay.qr.dto.GuestManualCheckInRequestDto;
-import com.fairing.fairplay.qr.dto.GuestQrCheckInRequestDto;
-import com.fairing.fairplay.qr.dto.MemberManualCheckInRequestDto;
-import com.fairing.fairplay.qr.dto.MemberQrCheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.CheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.CheckInResponseDto;
+import com.fairing.fairplay.qr.dto.scan.GuestManualCheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.GuestQrCheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.MemberManualCheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.MemberQrCheckInRequestDto;
 import com.fairing.fairplay.qr.dto.QrTicketRequestDto;
 import com.fairing.fairplay.qr.entity.QrActionCode;
 import com.fairing.fairplay.qr.entity.QrCheckStatusCode;
 import com.fairing.fairplay.qr.entity.QrTicket;
-import com.fairing.fairplay.qr.repository.QrTicketRepository;
 import com.fairing.fairplay.qr.util.CodeValidator;
 import java.time.LocalDateTime;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 /*
@@ -37,10 +32,10 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class QrTicketEntryService {
 
-  private final QrTicketRepository qrTicketRepository;
-  private final AttendeeRepository attendeeRepository;
   private final CodeValidator codeValidator; // 검증 로직 분리
+  private final QrTicketVerificationService qrTicketVerificationService;
   private final QrLogService qrLogService;
+  private final QrTicketAttendeeService qrTicketAttendeeService;
   private final QrEntryValidateService qrEntryValidateService;
 
   private static final String QR = "QR";
@@ -49,7 +44,7 @@ public class QrTicketEntryService {
   // 회원 QR 코드 체크인
   public CheckInResponseDto checkIn(MemberQrCheckInRequestDto dto, CustomUserDetails userDetails) {
     // 예약자 조회
-    Attendee attendee = findAttendee(dto.getReservationId(), null);
+    Attendee attendee = qrTicketAttendeeService.findAttendee(dto.getReservationId(), null);
     CheckInRequestDto checkInRequestDto = CheckInRequestDto.builder()
         .attendee(attendee)
         .requireUserMatch(true)
@@ -66,7 +61,7 @@ public class QrTicketEntryService {
   public CheckInResponseDto checkIn(MemberManualCheckInRequestDto dto,
       CustomUserDetails userDetails) {
     // 예약자 조회
-    Attendee attendee = findAttendee(dto.getReservationId(), null);
+    Attendee attendee = qrTicketAttendeeService.findAttendee(dto.getReservationId(), null);
     CheckInRequestDto checkInRequestDto = CheckInRequestDto.builder()
         .attendee(attendee)
         .requireUserMatch(true)
@@ -85,7 +80,7 @@ public class QrTicketEntryService {
     QrTicketRequestDto qrLinkTokenInfo = codeValidator.decodeToDto(dto.getQrLinkToken());
 
     // 예약자 조회
-    Attendee attendee = findAttendee(null, qrLinkTokenInfo.getAttendeeId());
+    Attendee attendee = qrTicketAttendeeService.findAttendee(null, qrLinkTokenInfo.getAttendeeId());
     CheckInRequestDto checkInRequestDto = CheckInRequestDto.builder()
         .attendee(attendee)
         .requireUserMatch(false)
@@ -104,7 +99,7 @@ public class QrTicketEntryService {
     QrTicketRequestDto qrLinkTokenInfo = codeValidator.decodeToDto(dto.getQrLinkToken());
 
     // 예약자 조회
-    Attendee attendee = findAttendee(null, qrLinkTokenInfo.getAttendeeId());
+    Attendee attendee = qrTicketAttendeeService.findAttendee(null, qrLinkTokenInfo.getAttendeeId());
     CheckInRequestDto checkInRequestDto = CheckInRequestDto.builder()
         .attendee(attendee)
         .requireUserMatch(false)
@@ -118,81 +113,15 @@ public class QrTicketEntryService {
   }
 
   /**
-   * Attendee 조회 (회원은 reservationId, 비회원은 attendeeId로)
-   */
-  private Attendee findAttendee(Long reservationId, Long attendeeId) {
-    if (reservationId != null) {
-      return findAttendeeByReservation(reservationId);
-    }
-    if (attendeeId != null) {
-      return findAttendeeByAttendee(attendeeId);
-    }
-    throw new CustomException(HttpStatus.BAD_REQUEST, "reservationId 또는 attendeeId 중 하나는 필수입니다.");
-  }
-
-  // 회원 attendee 조회
-  private Attendee findAttendeeByReservation(Long reservationId) {
-    return attendeeRepository.findByReservation_ReservationIdAndAttendeeTypeCode_Code(
-            reservationId, AttendeeTypeCode.PRIMARY)
-        .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "참석자를 조회하지 못했습니다."));
-  }
-
-  // 비회원 attendee 조회
-  private Attendee findAttendeeByAttendee(Long attendeeId) {
-    return attendeeRepository.findById(attendeeId)
-        .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "참석자를 조회하지 못했습니다."));
-  }
-
-  /**
-   * 로그인 사용자와 참석자 일치 여부 확인
-   */
-  private void validateUserMatch(Attendee attendee, CustomUserDetails userDetails) {
-    if (!attendee.getReservation().getUser().getUserId().equals(userDetails.getUserId())) {
-      throw new CustomException(HttpStatus.FORBIDDEN, "참석자와 로그인한 사용자가 일치하지 않습니다.");
-    }
-  }
-
-  /**
-   * QR 티켓 조회
-   */
-  private QrTicket findQrTicket(Attendee attendee) {
-    return qrTicketRepository.findByAttendee(attendee)
-        .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, "예약된 QR 티켓을 찾지 못했습니다."));
-  }
-
-  /**
-   * 코드 검증
-   */
-  // QR 코드 검증
-  private void verifyQrCode(QrTicket qrTicket, String qrCode) {
-    if (!qrTicket.getActive()) {
-      throw new CustomException(HttpStatus.BAD_REQUEST, "만료된 QR 티켓입니다.");
-    }
-
-    if (!qrTicket.getQrCode().equals(qrCode)) {
-      throw new CustomException(HttpStatus.UNAUTHORIZED, "참석자 정보와 일치하지 않습니다.");
-    }
-  }
-
-  // 수동 코드 검증
-  private void verifyManualCode(QrTicket qrTicket, String manualCode) {
-    codeValidator.validateManualCode(manualCode);
-
-    if (!qrTicket.getManualCode().equals(manualCode)) {
-      throw new CustomException(HttpStatus.UNAUTHORIZED, "참석자 정보와 일치하지 않습니다.");
-    }
-  }
-
-  /**
    * 체크인 공통 로직
    */
   private CheckInResponseDto processCheckInCommon(CheckInRequestDto dto) {
     if (dto.isRequireUserMatch()) {
       // 회원, 비회원 여부에 따라 참석자=로그인한 사용자 여부 조회
-      validateUserMatch(dto.getAttendee(), dto.getUserDetails());
+      qrTicketVerificationService.validateUserMatch(dto.getAttendee(), dto.getUserDetails());
     }
     // QR 티켓 조회
-    QrTicket qrTicket = findQrTicket(dto.getAttendee());
+    QrTicket qrTicket = qrTicketVerificationService.findQrTicket(dto.getAttendee());
     // QrActionCode 검토
     QrActionCode qrActionCode = qrEntryValidateService.validateQrActionCode(QrActionCode.SCANNED);
     // 코드 스캔 기록
@@ -205,9 +134,9 @@ public class QrTicketEntryService {
     qrEntryValidateService.preventDuplicateScan(qrTicket, entryType);
     // 코드 비교
     if (QR.equals(dto.getCodeType())) {
-      verifyQrCode(qrTicket, dto.getCodeValue());
+      qrTicketVerificationService.verifyQrCode(qrTicket, dto.getCodeValue());
     } else {
-      verifyManualCode(qrTicket, dto.getCodeValue());
+      qrTicketVerificationService.verifyManualCode(qrTicket, dto.getCodeValue());
     }
     // QR 티켓 처리
     LocalDateTime checkInTime = processCheckIn(qrTicket, dto.getQrActionCode(), entryType);
