+++ src/main/java/com/fairing/fairplay/qr/service/QrTicketExitService.java
@@ -1,0 +1,10 @@
+package com.fairing.fairplay.qr.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class QrTicketExitService {
+
+}
