+++ src/main/java/com/fairing/fairplay/qr/service/QrTicketService.java
@@ -1,11 +1,12 @@
 package com.fairing.fairplay.qr.service;
 
 import com.fairing.fairplay.core.security.CustomUserDetails;
-import com.fairing.fairplay.qr.dto.CheckInResponseDto;
-import com.fairing.fairplay.qr.dto.GuestManualCheckInRequestDto;
-import com.fairing.fairplay.qr.dto.GuestQrCheckInRequestDto;
-import com.fairing.fairplay.qr.dto.MemberManualCheckInRequestDto;
-import com.fairing.fairplay.qr.dto.MemberQrCheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.CheckInResponseDto;
+import com.fairing.fairplay.qr.dto.scan.CheckOutResponseDto;
+import com.fairing.fairplay.qr.dto.scan.GuestManualCheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.GuestQrCheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.MemberManualCheckInRequestDto;
+import com.fairing.fairplay.qr.dto.scan.MemberQrCheckInRequestDto;
 import com.fairing.fairplay.qr.dto.QrTicketReissueRequestDto;
 import com.fairing.fairplay.qr.dto.QrTicketReissueResponseDto;
 import com.fairing.fairplay.qr.dto.QrTicketRequestDto;
@@ -24,7 +25,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class QrTicketService {
 
   private final QrTicketIssueService qrTicketIssueService;
-  private final QrTicketEntryService qrTicketEntryService;
+  private final EntryExitService entryExitService;
 
   // 회원 QR 티켓 조회 -> 마이페이지에서 조회
   @Transactional
@@ -60,40 +61,5 @@ public class QrTicketService {
   @Transactional
   public QrTicketReissueResponseDto reissueAdminQrTicket(QrTicketReissueRequestDto dto) {
     return qrTicketIssueService.reissueAdminQrTicket(dto);
-  }
-
-  /*
-   * 체크인
-   * 1. 회원+QR
-   * 2. 회원+수동
-   * 3. 비회원+QR
-   * 4. 비회원+수동
-   *
-   */
-
-  // 체크인 1
-  @Transactional
-  public CheckInResponseDto checkInWithQrByMember(MemberQrCheckInRequestDto dto,
-      CustomUserDetails userDetails) {
-    return qrTicketEntryService.checkIn(dto, userDetails);
-  }
-
-  // 체크인 2
-  @Transactional
-  public CheckInResponseDto checkInWithManualByMember(MemberManualCheckInRequestDto dto,
-      CustomUserDetails userDetails) {
-    return qrTicketEntryService.checkIn(dto, userDetails);
-  }
-
-  // 체크인 3
-  @Transactional
-  public CheckInResponseDto checkInWithQrByGuest(GuestQrCheckInRequestDto dto) {
-    return qrTicketEntryService.checkIn(dto);
-  }
-
-  // 체크인 4
-  @Transactional
-  public CheckInResponseDto checkInWithManualByGuest(GuestManualCheckInRequestDto dto) {
-    return qrTicketEntryService.checkIn(dto);
   }
 }
